Move test_fixtures and render_context to separate files
